# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: ScanRay Sonar API
  description: |
    Single-pane-of-visibility API for the ScanRay Sonar platform.
    Phase 1 surface: authentication, sites, users, agents, appliances,
    plus health and version endpoints.

    All non-auth endpoints require `Authorization: Bearer <jwt>`.
    Mutating endpoints additionally require RBAC role >= the level
    documented per-operation.
  version: "2026.5.5.9"
  license:
    name: MIT
servers:
  - url: /api/v1
    description: Versioned API root

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---- Reusable schemas -------------------------------------------------
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        totp: { type: string, description: "Required if MFA enrolled" }
    LoginResponse:
      type: object
      required: [accessToken, refreshToken, expiresAt, user]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresAt: { type: string, format: date-time }
        mfaRequired: { type: boolean, default: false }
        user: { $ref: "#/components/schemas/User" }
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /auth/login:
    post:
      summary: Exchange email/password (+optional TOTP) for tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Login OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/refresh:
    post:
      summary: Refresh an access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/me:
    get:
      summary: Current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /sites:
    get:
      summary: List sites visible to the caller
      responses:
        "200":
          description: Sites
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Site" }
    post:
      summary: Create a new site (superadmin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSiteRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Site" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /users:
    get:
      summary: List users (superadmin)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      summary: Create a user (superadmin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /agents:
    get:
      summary: List agents (filter by site)
      parameters:
        - in: query
          name: siteId
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Agent" }
  /users/{id}:
    patch:
      summary: Update a user (superadmin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      summary: Permanently delete a user (superadmin)
      description: |
        Hard-deletes the user and cascade-removes their site memberships
        and API keys. Audit log entries created by the user are
        preserved (the actor_id becomes an orphan reference, intentional).
        Returns 400 with code `self_delete` when the caller targets their
        own account, and 400 with code `last_superadmin` when removing
        the user would leave zero active superadmins.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /topology:
    get:
      summary: Network topology graph (managed appliances + LLDP/CDP neighbors)
      description: |
        Builds a deduped node/edge graph from every appliance's
        last_snapshot. Managed nodes are real Appliance rows; foreign
        nodes are devices we only learned about via a neighbor's CDP
        or LLDP advertisement (id is `foreign:<sysname>`). Edges are
        keyed by an unordered node-pair so a link reported by both
        ends collapses into one edge with `protocol: "both"`.
      parameters:
        - in: query
          name: siteId
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Topology
          content:
            application/json:
              schema:
                type: object
                properties:
                  generatedAt: { type: string, format: date-time }
                  nodes:
                    type: array
                    items:
                      type: object
                      properties:
                        id:        { type: string }
                        kind:      { type: string, enum: [appliance, foreign] }
                        name:      { type: string }
                        label:     { type: string }
                        vendor:    { type: string }
                        model:     { type: string }
                        platform:  { type: string }
                        mgmtIp:    { type: string }
                        siteId:    { type: string, format: uuid }
                        status:    { type: string, enum: [up, degraded, down, unknown] }
                        lastSeenAt:  { type: string, format: date-time }
                        portsTotal:  { type: integer }
                        portsUp:     { type: integer }
                        uplinkCount: { type: integer }
                  edges:
                    type: array
                    items:
                      type: object
                      properties:
                        from:      { type: string }
                        to:        { type: string }
                        fromPort:  { type: string }
                        toPort:    { type: string }
                        protocol:  { type: string, enum: [lldp, cdp, both] }
                        operUp:    { type: boolean }
  /appliances:
    get:
      summary: List appliances (filter by site)
      parameters:
        - in: query
          name: siteId
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Appliances
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Appliance" }
    post:
      summary: Create a new appliance (siteadmin+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateApplianceRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appliance" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /appliances/{id}:
    get:
      summary: Get a single appliance with its latest SNMP snapshot
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Appliance
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Appliance"
                  - type: object
                    properties:
                      sysDescr: { type: string, nullable: true }
                      lastSnapshotAt: { type: string, format: date-time, nullable: true }
                      lastSnapshot:
                        type: object
                        description: |
                          Verbatim SNMP snapshot the poller stored. Contains
                          system, chassis, interfaces[], entities[] and lldp[]
                          arrays — see internal/snmp/snapshot.go for the full
                          schema.
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: Remove an appliance (siteadmin+)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /agents/{id}:
    get:
      summary: Get a single agent with its latest snapshot
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Agent"
                  - type: object
                    properties:
                      lastMetrics:
                        type: object
                        description: |
                          Verbatim last snapshot the probe sent (Snapshot v2). Includes
                          a `conversations` array of aggregated active TCP/UDP peer
                          pairs (direction, remoteIp, remoteHost, remotePort, localPort,
                          state, pid, processName, count) for the "talking to" view.
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: Remove an agent (siteadmin+)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /agents/{id}/network:
    get:
      summary: Network bps samples + hourly byte totals
      description: |
        One sample per snapshot (60 s cadence) plus a 1-hour
        rollup of bytes-sent / bytes-received. Powers the
        "Network usage" + "Network usage (hourly)" charts on
        AgentDetail.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: range
          schema: { type: string, default: "24h" }
      responses:
        "200":
          description: Samples + hourly rollup
          content:
            application/json:
              schema:
                type: object
                properties:
                  agentId: { type: string, format: uuid }
                  range:   { type: string }
                  samples:
                    type: array
                    items:
                      type: object
                      properties:
                        time:   { type: string, format: date-time }
                        inBps:  { type: integer, nullable: true }
                        outBps: { type: integer, nullable: true }
                  hourly:
                    type: array
                    items:
                      type: object
                      properties:
                        hour:     { type: string, format: date-time }
                        inBytes:  { type: integer, nullable: true }
                        outBytes: { type: integer, nullable: true }
  /agents/{id}/latency:
    get:
      summary: ICMP latency samples per target
      description: |
        Samples produced by the probe's ICMP latency module, one row
        per (target, time). target='8.8.8.8' is the WAN target; target
        ='gateway' is the host's discovered default-route gateway.
        Older probes (schema < v4) return an empty array.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: range
          schema: { type: string, default: "24h" }
        - in: query
          name: target
          schema: { type: string, enum: ["both", "8.8.8.8", "gateway"], default: "both" }
      responses:
        "200":
          description: Samples
          content:
            application/json:
              schema:
                type: object
                properties:
                  agentId: { type: string, format: uuid }
                  range:   { type: string }
                  target:  { type: string }
                  samples:
                    type: array
                    items:
                      type: object
                      properties:
                        time:    { type: string, format: date-time }
                        target:  { type: string }
                        address: { type: string, nullable: true }
                        avgMs:   { type: number, nullable: true }
                        minMs:   { type: number, nullable: true }
                        maxMs:   { type: number, nullable: true }
                        lossPct: { type: number, nullable: true }
...
  /agents/enrollment-tokens:
    get:
      summary: List outstanding enrollment tokens (siteadmin+)
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EnrollmentToken" }
    post:
      summary: Issue a new enrollment token (siteadmin+)
      description: |
        Returns a plaintext token in the response — capture it
        immediately, the API only stores its sha256 hash.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEnrollmentTokenRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NewEnrollmentToken" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /agents/enrollment-tokens/{id}:
    delete:
      summary: Revoke an enrollment token (siteadmin+)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/NotFound" }
  /agents/enroll:
    post:
      summary: Enroll an agent using a one-time token (unauthenticated)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentEnrollRequest" }
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentEnrollResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /probe/download/{os}/{arch}:
    get:
      summary: Download a probe binary (unauthenticated)
      security: []
      parameters:
        - in: path
          name: os
          required: true
          schema: { type: string, enum: [linux, windows] }
        - in: path
          name: arch
          required: true
          schema: { type: string, enum: [amd64, arm64] }
      responses:
        "200":
          description: binary
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /sites/{id}/network-map:
    get:
      summary: Site-scoped topology graph
      tags: [Topology]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: topology graph
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
...
  /discovery/networks:
    get:
      summary: Per-site discovery subnets and device counts
      tags: [Discovery]
      parameters:
        - in: query
          name: siteId
          schema: { type: string, format: uuid }
        - in: query
          name: collectorId
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: networks
          content:
            application/json:
              schema:
                type: array
                items: { type: object, additionalProperties: true }
...